feat: rename page
